pygrass: replace deprecated numpy.bool with bool (#2949)
* another deprecated alias replaced in GUI scatterplot

+++ gui/wxpython/iscatt/plots.py
@@ -929,7 +929,7 @@ class ModestImage(mi.AxesImage):
         self._full_res = A
         self._A = A
 
-        if self._A.dtype != np.uint8 and not np.can_cast(self._A.dtype, np.float):
+        if self._A.dtype != np.uint8 and not np.can_cast(self._A.dtype, float):
             raise TypeError("Image data can not convert to float")
 
         if self._A.ndim not in (2, 3) or (
